requirements.txt
python-telegram-bot==22.7
telethon==1.34.0
python-dotenv==1.0.0
aiosqlite==0.19.0
